Move service::Session to its own componet
